+++ Rototo/Includes/Primitive.h
@@ -37,7 +37,7 @@ public:
 	void SetAngles(int _start,int _end);
 	void SetSize(int _w,int _h);
   void SetPosition(int _x,int _y);
-	void SetScale(double _xFactor,double _yFactor);
+  void SetScale(float _xFactor, float _yFactor);
 	void Render();
 	void SetColor(unsigned char _r=255,unsigned char _g=255,unsigned char _b=255,unsigned char _a=255);
